Fix exception hierarchy
In cases where a new redhat OWNERS file is added but the chart doesn't
contain the redhat- prefix, the workflow would fail without creating a
comment for the user due to trying to download the Submission artifact.

This artifact is missing (this is expected), because there was an issue
during the parsing of the modified file in 'validate-submission'. This
step should created a GitHub output 'submission_file_present'. This was
not happening in that particular case due to 'ChartError' (exception
raised when RedHat chart is missing the redhat- prefix) not being a
SubmissionError exception.

Fix #465

Signed-off-by: Matthias Goerens <mgoerens@redhat.com>