Prioritize long running behave feature (#469)
The HC-16 feature takes more than 90 minutes to complete. To optimize
the duration of end to end tests, this commits ensures that it is
starting first.

The order in which the features are placed in the array is not
deterministic when using the find command i.e. the features are ran in a
random order.

The sort_by jq filter places the element starting with "HC-16" string at
the end of the array, requiring to then reverse the array to place it at
the front.

Fix #407

Signed-off-by: Matthias Goerens <mgoerens@redhat.com>