Fix error message for PRs with OWNERS and chart (#479)
PRs containing a chart and an OWNERS file were getting the wrong error
message. This was due to the check for duplicate charts, not taking into
account that an OWNERS file doens't relate to a specific version, and
thus considering the OWNERS file to be for a different chart.

Specific tests have been added to cover this case.

Fix #477

Signed-off-by: Matthias Goerens <mgoerens@redhat.com>